+++ jdaviz/configs/mosviz/plugins/parsers.py
@@ -755,21 +755,25 @@ def mos_niriss_parser(app, data_dir):
             print(f"Loading: {flabel} sources")
 
             with fits.open(fname, memmap=False) as temp:
-                # treat all HDUs missing SRCTYPE attribute as extended sources
-                # TODO: Remove this once valid SRCTYPE values are present in all headers
-                for hdu in temp:
-                    if ("SRCTYPE" in hdu.header
-                            and (hdu.header["SRCTYPE"] in ("POINT", "EXTENDED"))):
-                        pass
-                    else:
-                        hdu.header["SRCTYPE"] = "EXTENDED"
+                # Filter out HDUs we care about
+                source_ids_to_filter = cat_id_dict.keys()
+                filtered_hdul = fits.HDUList([hdu for hdu in temp if (
+                    (hdu.name in ('PRIMARY', 'ASDF')) or
+                    (hdu.header.get('SOURCEID', None) in source_ids_to_filter))])
 
-                # read all HDUs with SpectrumList, then only keep those that
-                # correspond with sources in catalog
-                # (read() is slow... would also LOVE to do this in one step!)
-                specs = SpectrumList.read(temp, format="JWST x1d multi")
-                specs_cut = [sp for sp in specs if sp.meta['header']['SOURCEID']
-                             in cat_id_dict.keys()]
+                # SRCTYPE is required for the specutils JWST x1d reader. The reader will
+                # force this to POINT if not set. Under known cases, this field will be set
+                # for NIRISS programs; if it's not, something's gone wrong. Catch this
+                # warning and reraise as an error to warn users.
+                try:
+                    with warnings.catch_warnings():
+                        warnings.filterwarnings("error",
+                                                category=UserWarning,
+                                                message=".*SRCTYPE is missing or UNKNOWN*")
+                        specs = SpectrumList.read(filtered_hdul, format="JWST x1d multi")
+                except UserWarning as e:
+                    raise KeyError(f"The SRCTYPE keyword in the header of file {fname}"
+                                   "is not populated (expected values: EXTENDED or POINT)") from e
 
                 filter_name = fits.getheader(fname, ext=0).get('PUPIL')
 
@@ -777,7 +781,7 @@ def mos_niriss_parser(app, data_dir):
                 orientation = flabel.split()[-1]
 
                 # update 1D labels and standardize metadata for table viewer
-                for sp in specs_cut:
+                for sp in specs:
                     if (
                         sp.meta['header']['SPORDER'] == 1
                         and sp.meta['header']['EXTNAME'] == 'EXTRACT1D'